Clean up easy leaks + prep for allocating Diffs
The tests currently create Diffs with constant/static text, which
causes a panic on any attempt to free said text.

This pass fixes a couple leaks, and lays a foundation for consistently
freeing any diff's .text field when no longer in use: this requires
that, as it will be in real programs, the diff text is allocated to
begin with.