Fix for diffCleanupMerge memory
I'll need to knock out all the tests in diffMergeSemantic so that
I can flip the rest of the diffCleanupMerge tests on, without trying
to free .rodata memory.

The real annoyance of the decision to punt on memory management is in
the tests, which all have to be rewritten.  Whoever you are, you could
at least have written the tests responsibly, and spared the person who
cleans up your sloppy code from a bunch of drudgery.