Code-bumming encodeUri
Reduces common subsequence to a pair of comparisons, and orders
test in approximate frequency of occurance (until we get to the
fribbles at the end, which are just in ascending order).

If I really cared I could reduce this to two u64 masks, a byte
mask, a switch on the four values of that result, and a shift-and-
compare.  Seems excessive though.  And I *like* excessive...