Add ownership handling to makePatch
Some of the functions which produce a patch, produce the Diffs as well.
Those functions can call makePatchInternal with the .own enum, and then
we don't have to copy the Diffs: the unused ones are freed, and the new
ones are allocated, while the edit diffs get moved to the Patch.

Then the original DiffList can be deinitialized without needing to also
free the Diff texts.