Add patch split and patch padding
Also has a sketch of an improved bitap which can potentially avoid
having to split patches in a great many cases.

This finishes the bulk of the work in translation.  It remains to
add the patch conversion functions, and decide what to do about the
to and from delta.  They aren't used, and I think the format could be
substantively improved.

Next up: get on the big monitor and do a proofread pass on all the code
added since the last tests, then translate tests from the suite up to a
point where this has total line coverage and no leaks.