Precalculate maximum space needed for makePatchInternal
I want to come up with a better all-around solution for this, the sheer
amount of copying going on is galling.

But this should give the first allocation enough room to accomodate
growing the text, considering that the allocator can reuse the memory,
although not in ReleaseSafe or Debug mode.