Check for lead byte in common suffix
Once I have a comprehensive test set for Unicode problems, I can circle
back and try to simplify some of these algorithms.

Specifically I don't like that commonSuffix is using a boolean to track
the prior status of bytes, it's a test in a hot loop that will actually
never trigger with ASCII, which is an important use case to support.

Ideally the Unicode code is slower than ASCII only in boundary checks.