Adds revamp of diffLine mode
Untested but ready for deployment.

In addition to using codepoints, this refactoring has a function taking
an iterator, which returns the lines.

This opens up the ability to stream from any iterator which the program
calls for.  This can include a tokenizer, for sharp semantic diffing of
source code files.